Extract serializers to own files
