+++ pylon-core/src/main/kotlin/io/github/pylonmc/pylon/core/persistence/datatypes/BlockPositionPersistentDataType.kt
@@ -1,0 +1,41 @@
+package io.github.pylonmc.pylon.core.persistence.datatypes
+
+import io.github.pylonmc.pylon.core.block.BlockPosition
+import io.github.pylonmc.pylon.core.pluginInstance
+import org.bukkit.NamespacedKey
+import org.bukkit.persistence.PersistentDataAdapterContext
+import org.bukkit.persistence.PersistentDataContainer
+import org.bukkit.persistence.PersistentDataType
+
+object BlockPositionPersistentDataType : PersistentDataType<PersistentDataContainer, BlockPosition> {
+    val worldKey = NamespacedKey(pluginInstance, "world")
+    val xKey = NamespacedKey(pluginInstance, "x")
+    val yKey = NamespacedKey(pluginInstance, "y")
+    val zKey = NamespacedKey(pluginInstance, "z")
+
+    override fun getPrimitiveType(): Class<PersistentDataContainer> = PersistentDataContainer::class.java
+
+    override fun getComplexType(): Class<BlockPosition> = BlockPosition::class.java
+
+    override fun fromPrimitive(
+        primitive: PersistentDataContainer,
+        context: PersistentDataAdapterContext
+    ): BlockPosition {
+        val x = primitive.get(xKey, PersistentDataType.INTEGER)!!
+        val y = primitive.get(yKey, PersistentDataType.INTEGER)!!
+        val z = primitive.get(zKey, PersistentDataType.INTEGER)!!
+        val world = primitive.get(worldKey, PylonSerializers.WORLD)
+        return BlockPosition(world, x, y, z)
+    }
+
+    override fun toPrimitive(complex: BlockPosition, context: PersistentDataAdapterContext): PersistentDataContainer {
+        val pdc = context.newPersistentDataContainer()
+        pdc.set(xKey, PersistentDataType.INTEGER, complex.x)
+        pdc.set(yKey, PersistentDataType.INTEGER, complex.y)
+        pdc.set(zKey, PersistentDataType.INTEGER, complex.z)
+        if (complex.world != null) {
+            pdc.set(worldKey, PylonSerializers.WORLD, complex.world!!)
+        }
+        return pdc
+    }
+}
