+++ pylon-core/src/main/kotlin/io/github/pylonmc/pylon/core/persistence/datatypes/CharPersistentDataType.kt
@@ -1,0 +1,22 @@
+package io.github.pylonmc.pylon.core.persistence.datatypes
+
+import org.bukkit.persistence.PersistentDataAdapterContext
+import org.bukkit.persistence.PersistentDataType
+import java.nio.ByteBuffer
+
+object CharPersistentDataType : PersistentDataType<ByteArray, Char> {
+    override fun getPrimitiveType(): Class<ByteArray> = ByteArray::class.java
+
+    override fun getComplexType(): Class<Char> = Char::class.java
+
+    override fun fromPrimitive(primitive: ByteArray, context: PersistentDataAdapterContext): Char {
+        val buffer = ByteBuffer.wrap(primitive)
+        return buffer.getChar()
+    }
+
+    override fun toPrimitive(complex: Char, context: PersistentDataAdapterContext): ByteArray {
+        val buffer = ByteBuffer.allocate(Char.SIZE_BYTES)
+        buffer.putChar(complex)
+        return buffer.array()
+    }
+}
