+++ pylon-core/src/main/kotlin/io/github/pylonmc/pylon/core/persistence/datatypes/InventoryPersistentDataType.kt
@@ -6,21 +6,6 @@ import org.bukkit.inventory.ItemStack
 import org.bukkit.persistence.PersistentDataAdapterContext
 import org.bukkit.persistence.PersistentDataType
 
-object ItemStackPersistentDataType : PersistentDataType<ByteArray, ItemStack> {
-
-    override fun getPrimitiveType(): Class<ByteArray> = ByteArray::class.java
-
-    override fun getComplexType(): Class<ItemStack> = ItemStack::class.java
-
-    override fun toPrimitive(complex: ItemStack, context: PersistentDataAdapterContext): ByteArray {
-        return complex.serializeAsBytes()
-    }
-
-    override fun fromPrimitive(primitive: ByteArray, context: PersistentDataAdapterContext): ItemStack {
-        return ItemStack.deserializeBytes(primitive)
-    }
-}
-
 object InventoryPersistentDataType : PersistentDataType<ByteArray, Inventory> {
 
     override fun getPrimitiveType(): Class<ByteArray> = ByteArray::class.java
