+++ pylon-core/src/main/kotlin/io/github/pylonmc/pylon/core/persistence/datatypes/ItemStackPersistentDataType.kt
@@ -1,0 +1,20 @@
+package io.github.pylonmc.pylon.core.persistence.datatypes
+
+import org.bukkit.inventory.ItemStack
+import org.bukkit.persistence.PersistentDataAdapterContext
+import org.bukkit.persistence.PersistentDataType
+
+object ItemStackPersistentDataType : PersistentDataType<ByteArray, ItemStack> {
+
+    override fun getPrimitiveType(): Class<ByteArray> = ByteArray::class.java
+
+    override fun getComplexType(): Class<ItemStack> = ItemStack::class.java
+
+    override fun toPrimitive(complex: ItemStack, context: PersistentDataAdapterContext): ByteArray {
+        return complex.serializeAsBytes()
+    }
+
+    override fun fromPrimitive(primitive: ByteArray, context: PersistentDataAdapterContext): ItemStack {
+        return ItemStack.deserializeBytes(primitive)
+    }
+}
