+++ pylon-core/src/main/kotlin/io/github/pylonmc/pylon/core/persistence/datatypes/PylonSerializers.kt
@@ -1,8 +1,6 @@
 package io.github.pylonmc.pylon.core.persistence.datatypes
 
-import org.bukkit.persistence.PersistentDataAdapterContext
 import org.bukkit.persistence.PersistentDataType
-import java.nio.ByteBuffer
 
 object PylonSerializers {
     @JvmField
@@ -82,48 +80,4 @@ object PylonSerializers {
 
     @JvmField
     val INVENTORY = InventoryPersistentDataType
-}
-
-object CharPersistentDataType : PersistentDataType<ByteArray, Char> {
-    override fun getPrimitiveType(): Class<ByteArray> = ByteArray::class.java
-
-    override fun getComplexType(): Class<Char> = Char::class.java
-
-    override fun fromPrimitive(primitive: ByteArray, context: PersistentDataAdapterContext): Char {
-        val buffer = ByteBuffer.wrap(primitive)
-        return buffer.getChar()
-    }
-
-    override fun toPrimitive(complex: Char, context: PersistentDataAdapterContext): ByteArray {
-        val buffer = ByteBuffer.allocate(Char.SIZE_BYTES)
-        buffer.putChar(complex)
-        return buffer.array()
-    }
-}
-
-class EnumPersistentDataType<E : Enum<E>>(val enumClass: Class<E>) : PersistentDataType<String, E> {
-
-    private val valueMap: Map<String, E> = enumClass.enumConstants.associateBy { it.name }
-
-    override fun getPrimitiveType(): Class<String> = String::class.java
-
-    override fun getComplexType(): Class<E> = enumClass
-
-    override fun toPrimitive(complex: E, context: PersistentDataAdapterContext): String {
-        return complex.name
-    }
-
-    override fun fromPrimitive(primitive: String, context: PersistentDataAdapterContext): E {
-        return valueMap[primitive] ?: throw IllegalArgumentException("Unknown enum value: $primitive")
-    }
-
-    companion object {
-        fun <E : Enum<E>> enumTypeFrom(enumClass: Class<E>): EnumPersistentDataType<E> {
-            return EnumPersistentDataType(enumClass)
-        }
-
-        inline fun <reified E : Enum<E>> enumTypeFrom(): EnumPersistentDataType<E> {
-            return EnumPersistentDataType(E::class.java)
-        }
-    }
 }
